pyproject.toml
[project]
name = "MyTools"
version = "0.1.0"
description = "bibi"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.15.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.hatch.build.targets.wheel]
packages = ["src/my_tools"]
